Day 15b complete
My implementation takes about 40s on my laptop--pretty slow, so if time
allows I'd like to optimize this solution a little. Decrease-key is slow
because it has to find the key; I think this could be improved at the
cost of memory by maintaining a reverse-index mapping key to its
position in the heap--but maybe there are other data structures (or
other heaps) that address this problem.